+++ src/main/java/dev/tonimatas/Main.java
@@ -1,9 +1,6 @@
 package dev.tonimatas;
 
-import dev.tonimatas.config.BankData;
-import dev.tonimatas.config.BotConfig;
-import dev.tonimatas.config.ExtraData;
-import dev.tonimatas.config.JsonFile;
+import dev.tonimatas.config.*;
 import dev.tonimatas.listeners.*;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
@@ -23,23 +20,17 @@ public class Main {
     private static final Logger LOGGER = LoggerFactory.getLogger(Main.class);
 
     public static void main(String[] args) {
-        BotConfig bot = JsonFile.loadOrCreate(BotConfig.class, "bot.json");
-        BankData bankData = JsonFile.loadOrCreate(BankData.class, "data/bank.json");
-        ExtraData extraData = JsonFile.loadOrCreate(ExtraData.class, "data/extra.json");
-
-        JDA jda = JDABuilder.createDefault(bot.token)
+        JDA jda = JDABuilder.createDefault(BotFiles.CONFIG.token)
                 .enableIntents(Arrays.stream(GatewayIntent.values()).toList())
                 .setMemberCachePolicy(MemberCachePolicy.ALL)
                 .setAutoReconnect(true)
                 .build();
 
-        jda.addEventListener(new RouletteListener(jda, bankData),
-                new SlashCommandListener(),
+        jda.addEventListener(new SlashCommandListener(),
                 new AutoRoleListener(),
-                new CountListener(extraData),
+                new CountListener(),
                 new JoinLeaveMessageListener(),
-                new TemporalChannelListener(),
-                new BankListener(bankData)
+                new TemporalChannelListener()
         );
 
         jda.updateCommands()
