+++ src/main/java/dev/tonimatas/config/BotFiles.java
@@ -1,0 +1,7 @@
+package dev.tonimatas.config;
+
+public class BotFiles {
+    public static BotConfig CONFIG = JsonFile.loadOrCreate(BotConfig.class, "bot.json");
+    public static BankData BANK = JsonFile.loadOrCreate(BankData.class, "data/bank.json");
+    public static ExtraData EXTRA = JsonFile.loadOrCreate(ExtraData.class, "data/extra.json");
+}
