+++ src/main/java/dev/tonimatas/systems/roulette/bets/ColumnBet.java
@@ -1,4 +1,4 @@
-package dev.tonimatas.roulette.bets;
+package dev.tonimatas.systems.roulette.bets;
 
 public class ColumnBet extends Bet {
     private final String input;
