+++ src/main/java/dev/tonimatas/systems/roulette/bets/NumberBet.java
@@ -1,4 +1,4 @@
-package dev.tonimatas.roulette.bets;
+package dev.tonimatas.systems.roulette.bets;
 
 public class NumberBet extends Bet {
     private int input;
